Add multi dc feature for hostNetwork cluster only
Provide capability to enable multi dc feature on scyllacluster crd with the list of distant seeds.
Provide a mechanism to define an initCluster. The init cluster will be the first bootstraped cluster. It will not rely on multi dc seed (only using local ones) so it can bootstrap autinomously.
Other clusters will rely on multi dc seeds only at bootstrap.
When bootstrap is finished meaning all statefulsets are fully created with all members ready a bootstrap status set to finished will be added to the scyllacluster status.
When the finished bootstrap status is reached the cluster will not anymore rely on multi dc seeds (rely on local one only)